move this to models
