+++ src/Models/VarianceDissipationComputations/VarianceDissipationComputations.jl
@@ -1,6 +1,6 @@
 module VarianceDissipationComputations
  
-export VarianceDissipation, get_dissipation_fields
+export VarianceDissipation, flatten_dissipation_fields
 
 using Oceananigans.Grids: architecture
 using Oceananigans.Utils
@@ -25,11 +25,8 @@ using Oceananigans.Advection: _advective_tracer_flux_x,
 
 using Oceananigans: UpdateStateCallsite
 using Oceananigans.Operators: volume
-using Oceananigans.Simulations: Simulation
 using Oceananigans.Utils: IterationInterval, ConsecutiveIterations
 using KernelAbstractions: @kernel, @index
-
-import Oceananigans.Simulations: Callback
 
 struct VarianceDissipation{P, K, A, D, S, G} 
     advective_production :: P
@@ -135,31 +132,5 @@ include("advective_dissipation.jl")
 include("diffusive_dissipation.jl")
 include("compute_dissipation.jl")
 include("flatten_dissipation_fields.jl")
-
-# A VarianceDissipation object requires a `ConsecutiveIteration` schedule to make sure
-# that the computed fluxes are correctly used in the next time step.
-# Also, the `VarianceDissipation` object needs to be called on `UpdateStateStepCallsite` to be correct.
-function Callback(func::VarianceDissipation, schedule;
-                  parameters = nothing,
-                  callsite = UpdateStateCallsite())
-
-    if !(callsite isa UpdateStateCallsite)
-        @warn "VarianceDissipation callback must be called on UpdateStateCallsite. Changing `callsite` to `UpdateStateCallsite()`."
-        callsite = UpdateStateCallsite()
-    end
-
-    if schedule isa TimeInterval # Time step might change size between iterations invalidating `VarianceDissipation`
-        throw(ArgumentError("a VarianceDissipation computation must be executed on `IterationInterval`s. \n" *
-                            "The provided `TimeInterval` schedule is not supported"))
-    end
-
-    if !(schedule isa ConsecutiveIterations || schedule == IterationInterval(1))
-        @warn "VarianceDissipation callback must be called every Iteration or on `ConsecutiveIterations`. \n" * 
-              "Changing `schedule` to `ConsecutiveIterations(schedule)`."
-        schedule = ConsecutiveIterations(schedule)
-    end
-
-    return Callback(func, schedule, callsite, parameters)
-end
 
 end
